Apply memory translation on RPi4
RPi4 has a different mmio base to the one of RPi3.

Originally, the mmio_base in RPi3 can be covered in the first 1 GB, but
to RPi4, the mmio_base (0xFE000000) is not in this range.

Due to this problem, I add a new table to map the RPi4's mmio_base, and
the table is simplily placed after the kernel page table.